requirements.txt
smbus
time
RPi.GPIO
bitstring